Organized files to two new places gif_related, and other_tools
