+++ other_tools/click_and_crop.py
